dot net warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6689 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/NetHandler.cxx
@@ -243,7 +243,7 @@ RxStatus NetHandler::receive(size_t length) {
   RxStatus returnValue;
   if ((int)length <= tcplen)
     return ReadAll;
-  int size = recv(fd, tcpmsg + tcplen, length - tcplen, 0);
+  int size = recv(fd, tcpmsg + tcplen, (int)length - tcplen, 0);
   if (size > 0) {
     tcplen += size;
     if (tcplen == (int)length)
